Move `impl_array_newtype` to `internals`
Discussion #5331 suggested removing `bitcoin` as a direct dependency
could be a reasonable goal for the `p2p` crate. One conflict along the
way is the use of this macro in `bip152` to implement the array-like
traits for the short IDs. Because `bitcoin` depends on `internals`, as
does `p2p`, we can move this macro into `internals` to help detangle
`p2p` and `bitcoin`.

Moves the deprecated methods to the `impl` blocks.
